fix many warnings
darcs-hash:20080616090651-2fc9d-45c501495cf06acc11da3316e5c5a7e0f7e83fce.gz

+++ itkAttributeSelectionLabelMapFilter.h
@@ -126,7 +126,7 @@ public:
     
   void AddAttribute(  const AttributeValueType & attr )
     {
-    int size = m_AttributeSet.size();
+    unsigned long size = m_AttributeSet.size();
     m_AttributeSet.insert( attr );
     if( size != m_AttributeSet.size() )
       {
